Add e2e support for specifying XNNPACK thread count (#16)
This PR adds the `iree-compile` options `xnnpack-threads` that specifies the number of threads in the XNNPACK threadpool. The option is currently plumbed through to each xnnpack dialect op after legalization. In order to achieve the use of flags, this PR also makes some minor changes to the pass autogeneration of `legalize-xnnpack`.

Current limitation:
- The threadpool is created every time an op is called. This can be fixed by
  - storing the threadpool after it gets created for the first time in the `void* context` variable that gets passed to every executable plugin function
  - creating xnnpack dialect ops for creating the threadpool, which can then be inserted as part of the MLIR