Match a fully connected pattern with PDLL
There are C++ constraints for type checking, but still PDLL is much conciser
than the C++ pattern matching.

We still need to do the transpose handling to fully replace the C++ pattern code.

Note that without the pattern file, the normal C++ pattern matcher will be used,
so we don't need another flag to tell which pattern matcher to use.