Move stablehlo-to-xnnpack to input conversion preprocessing pipeline
IREE converts StableHLO to Linalg before calling the preprocessing
pass pipeline in the compiler plugins, preventing the
`stablehlo-to-xnnpack` pass from performing the rewrite at the right
time.

This commit moves the `stablehlo-to-xnnpack` pass to the input
conversion preprocessing pipeline to rewrite the graph before any
other input conversions are performed.