Convert kernel to unsigned before passing to fully_connected op
XNNPACK currently expects the input kernel tensor to be of type
`ui4`. This commit add the conversion to unsigned by manually XORing
the kernel tensor with the value 8.

Note: doing an `ui8->i4` conversion does not result in a `ui4` tensor,
and a `ui8->ui4` conversion is currently not handled by
IREE. Therefore, we manually convert to unsign by XORing.