Add transpose attribute to fully_connected op
This commit adds an boolean attribute to the `fully_connected` XNNPACK
op for specifying if the `kernel` tensor requires a transpose before
the fully connected computation is performed. The attribute is then
passed as the last argument in the `params` struct to the executable
plugin, so that XNNPACK can perform the tranpose of the weight.

Future work includes adding the ability for specifying whether to make
the transpose in stablehlo (as is done before this commit) or in
XNNPACK.