Restrict fully_connected op to 2d input and output (#27)
This commit restricts the fully connected op of XNNPACK to only take
2d inputs and produce a 2d output. This matches the semantics of the
op in the XNNPACK library. In order to keep the 3d input/output
support with batch-size of 1, reshape ops are inserted before and
after the op.

The insertion of reshape operations leads to a compiler error
regarding the hoisting of global variables. The errors turns out to be
due to poor handling of function calls by the IREE compiler. If an
`inline` pass is inserted right after reaching the `flow` dialect, all
errors regarding hoisting and consteval go away. Therefore, this
commit also changes the lit tests to include an `inline` pass after
reaching the `flow` level.

Note: currently there is no simple way to add passes after the `flow`
level from the compiler plugin, so the lit tests do it
manually. Future work will include moving this change to the pipeline
to the compiler plugin.