Add type check to getUnconvertedValue/getConvertedValue helpers
Sometimes the input to the `stablehlo.dot_general` can be the type
needed for the `fully_connected` op while at the same time being the
result of a `stablehlo.convert` op. In these cases, we want to keep
the input _after_ the `convert` rather than before.

This commit adds a type check to the helper functions that crawl up
and down the `stablehlo.convert` ops to make sure they stop if the
right type is already there.