Add cache support for fully_connected XNNPACK op (#29)
This commit adds a cache to the XNNPACK executable plugin to avoid
creating the same op multiple times. In order to know if two fully
connected XNNPACK ops are the same, we need to ensure that the
`kernel` parameter is the same. This information is propagated by the
new pass `SetIdAttributes`, which goes through each fully connected op
and assigns an ID to each unique `kernel`.